Use verbatim path with "load" command.
Prior to this commit "load /path/blah.f" didn't work because
"load" always added the ".f" extention to the provided input.
That's confusing - if a path to a file is given it should be
used as is. This commit removes appending .f to the load arg.

Note that the fallback, which is loading workload defs from
the files installed under /usr/local/... is not changed and
"load webserver" still works provided webserver.f  is under
/usr/local/share/filebench/workloads/ (or as installed).

Signed-off-by: Michal Ludvig <mludvig@logix.net.nz>